pyproject.toml
[tool.poetry]
name = "excel-translate"
version = "0.1.0"
description = "translate excel columns based on customized rules!"
authors = ["TRoYals <aozora00321@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "excel_translate"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.244"
jupyter = "^1.0.0"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.2"



[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
excel_translate = "main"